FindSciPetsc can now look for PETSc libs in directories with names other than petsc.
Added a check for the PETSC_FIND_VERSION variable which, if found, is
appended onto the directory name. So if, for instance, one were to do
set(PETSC_FIND_VERSION "3.3")
in CMakeLists.txt before the call to FindSciPetsc, then it would look
in "petsc3.3" (in serial) or "petsc3.3-par" (in parallel) for the
PETSc libraries.


git-svn-id: https://ice.txcorp.com/svnrepos/code/scimake/trunk@39 03405629-b146-45b9-a56f-558d5653c1c1