THIS IS JUST A CORRECTION FOR THE PREVIOUS COMMIT MESSAGE WHICH HAD A MISTAKE AND SHOULD INSTEAD STATE:
Also modified SciFindPackage by adding an optional parameter called ALLOW_LIBRARY_DUPLICATES which when set to TRUE allows for the list of libraries returned to include duplicates of libraries. If it is set to FALSE or not defined, duplicate libraries are taken out, as previous revisions did by default. This is done for possible circular dependencies; notably when building GPULib and linking it with MAGMA and Atlas.



git-svn-id: https://ice.txcorp.com/svnrepos/code/scimake/trunk@54 03405629-b146-45b9-a56f-558d5653c1c1