Refactoring unit tests under CMake.
There is now one universal macro, SciAddUnitTest(), which takes the 
following arguments:

NAME        = the name of this test (which may or may not be the same
              as the executable)
COMMAND     = test executable (typically same as NAME, but need not be)
SOURCES     = 1+ source files to be compiled
LIBS        = libraries needed to link test
ARGS        = arguments to test
EXEC_DIRS   = directories to add to PATH when executing test
RESULTS     = Files to be compared against golden results. If this
              var is empty, no comparisons will be done (but see
              STDOUT below)
RESULTS_DIR = directory which contains expected results
STDOUT_FILE = Name of file into which stdout should be captured. This
              file will be added to $RESULTS so it will be compared
              against expected output.
 
Some shenanigans were required because one cannot pass a list to cmake
via the -D argument. Once we switch to CMake 2.8.11, this will be
simlified a little, in that we can use the $<SEMICOLON> generator
instead of what I am currently using, ASCII 001 (ctrl-A).



git-svn-id: https://ice.txcorp.com/svnrepos/code/scimake/trunk@450 03405629-b146-45b9-a56f-558d5653c1c1