Further improvements in Intel MPI calculation
If you need the explicit MPI libraries, this goes through
the output of `mpiicc -show` and find_file/find_library to generate
a list of the the full path to the libraries.  Right now MPI_LIBRARY
contains the dynamic libraries, but this can be changed to static easily
if needed.  Not sure what Vorpal needs exactly but this is trivially
close.


git-svn-id: https://ice.txcorp.com/svnrepos/code/scimake/trunk@729 03405629-b146-45b9-a56f-558d5653c1c1